pyproject.toml
[project]
name = "sudoku-solver"
version = "0.1.0"
description = "Sudoku solver using computer vision"
requires-python = ">=3.13"
dependencies = [
    "opencv-contrib-python>=4.10.0",
    "pillow>=12.1.1",
    "pytesseract>=0.3.13",
    "ultralytics>=8.4.23",
    "numpy>=2.0.0",
    "imutils>=0.5.4",
    "scikit-image>=0.26.0",
    "py-sudoku>=2.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "python-multipart>=0.0.20",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "ipykernel>=6.29.0",
    "jupyter>=1.1.1",
]
browser = [
    "playwright>=1.58.0",
]
local = [
    "label-studio>=1.23.0",
    "requests>=2.32.0",
]
